start layout for test

+++ app/Http/Controllers/Dashboard/AdminController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Dashboard;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Models\Role;
+use http\Env\Response;
 use Illuminate\Contracts\Validation\Rule;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -16,38 +17,40 @@ class AdminController extends Controller
 
 
     // метод получения списка пользователей через api в json, либо конкретного пользователя, если прилетает параметр
-    public function usersList(Request $request) {
+    public function usersList(Request $request)
+    {
 
-            $id = $request->all();
+        $id = $request->all();
 
-            if(empty($id)) {
-                $users = User::get(['id', 'name', 'first_name', 'last_name', 'email', 'middle_name']);
+        if (empty($id)) {
+            $users = User::get(['id', 'name', 'first_name', 'last_name', 'email', 'middle_name']);
 
-                foreach ($users as $user) {
-                    $role = $user->role()->first()->role_name;
-                    $user->role = $role;
-
-                }
-                return response()->json([
-                    'status'=>'success',
-                    'users'=>$users->toArray(),
-                ]);
-            } else {
-                $user = User::find($id)->first();
+            foreach ($users as $user) {
                 $role = $user->role()->first()->role_name;
                 $user->role = $role;
-                return  response()->json([
-                    'status'=>'success',
-                    'user'=>$user->toArray()
-                ],200);
+
             }
+            return response()->json([
+                'status' => 'success',
+                'users' => $users->toArray(),
+            ]);
+        } else {
+            $user = User::find($id)->first();
+            $role = $user->role()->first()->role_name;
+            $user->role = $role;
+            return response()->json([
+                'status' => 'success',
+                'user' => $user->toArray()
+            ], 200);
+        }
     }
 
     // метод обновления информации пользователя при редактировании
-    public function updateUser(Request $request) {
+    public function updateUser(Request $request)
+    {
         $val = Validator::make($request->all(), [
-        // в данный момент запрещено изменять login из панели администратора
-        //           'name'=> 'required|unique:users',
+            // в данный момент запрещено изменять login из панели администратора
+            //           'name'=> 'required|unique:users',
         ]);
 
         if ($val->fails()) {
@@ -63,21 +66,31 @@ class AdminController extends Controller
         $user->last_name = $request->last_name;
         $user->middle_name = $request->middle_name;
         $user->save();
-        $user->role()->update(['user_id'=>$request->id, 'role_name'=>$request->role]);
+        $user->role()->update(['user_id' => $request->id, 'role_name' => $request->role]);
 
         return response()->json([
-            'status' =>'success',
+            'status' => 'success',
             'user' => $user
         ], 200);
     }
+
     // метод удаления пользователя
-    public function deleteUser(Request $request) {
-        $user = User::find($request->id)->first();
-        $user->role()->remove();
+    public function deleteUser(User $user, Request $request)
+    {
+        //$user = User::find($request->id);
+        if ($user->name == 'admin') {
+            return response()->json([
+                'status' => 'error',
+                'error' => 'you can\'t delete admin',
+            ], 422);
+        }
+        $user->role()->delete();
         $user->delete();
 
         return response()->json([
-            'status'=>'success'
+            'status' => 'success',
+            'request' => $request->id,
+            'user' => $user
         ], 200);
     }
 }
